// src/api/schema.d.ts
export interface components {
    schemas: {
        AdminCreate: {
            readonly id: number;
            first_name: string;
            last_name: string;
            readonly role: components["schemas"]["Role"];
            role_id: number;
            phone: string;
            password: string;
            /** Format: date-time */
            readonly updated_at: string;
            /** Format: date-time */
            readonly created_at: string;
            /**
             * Oxirgi kirish
             * Format: date-time
             */
            last_login?: string | null;
        };
// ...
        Role: {
            readonly id: number;
            code: string;
            name: string;
        };
// ...
        UserSalary: {
            readonly id: number;
            readonly user: components["schemas"]["User"];
            user_id: number;
            /** Format: decimal */
            amount: string;
            /** Format: date */
            month?: string | null;
            /** Format: date-time */
            readonly created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}